Added authorization and authentication for endpoints

+++ src/main/java/de/playground/controller/UserController.java
@@ -1,31 +1,29 @@
 package de.playground.controller;
 
 import de.playground.dtos.User;
+import de.playground.security.IAccessUtils;
 import de.playground.services.IUserService;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/users")
 @AllArgsConstructor
+@Slf4j
 public class UserController {
     private final IUserService userService;
-
-    @PostMapping()
-    public ResponseEntity<?> createUser(@RequestBody User user) {
-        if (user == null || user.IsInvalid()) {
-            return ResponseEntity.badRequest().body("User is invalid.");
-        }
-
-        this.userService.createUser(user);
-        return ResponseEntity.ok().build();
-    }
+    private final IAccessUtils accessService;
 
     @PutMapping()
     public ResponseEntity<?> updateUser(@RequestBody User user) {
         if (user == null || user.IsInvalid()) {
             return ResponseEntity.badRequest().body("User is invalid.");
+        }
+
+        if (!accessService.authorizeUser(user.getUsername())) {
+            return ResponseEntity.status(403).build();
         }
 
         this.userService.updateUser(user);
@@ -38,6 +36,11 @@ public class UserController {
             return ResponseEntity.badRequest().body("Username is invalid.");
         }
 
+        if (!accessService.authorizeUser(username)) {
+            return ResponseEntity.status(403).build();
+        }
+
+
         this.userService.deleteUser(username);
         return ResponseEntity.ok().build();
     }
@@ -46,6 +49,10 @@ public class UserController {
     public ResponseEntity<?> readUser(@PathVariable String username) {
         if (username == null || username.trim().isEmpty()) {
             return ResponseEntity.badRequest().body("Username is invalid.");
+        }
+
+        if (!accessService.authorizeUser(username)) {
+            return ResponseEntity.status(403).build();
         }
 
         var user = this.userService.readUser(username);
